+++ CHANGELOG.md
@@ -5,10 +5,14 @@
 **New Features**
 
 * Add `getNetworkClient` method (`@colony/colony-js-client`)
+* Add `getFundingPot` and `getFundingPotCount` (`@colony/colony-js-client`)
+* Add temporary `makePayment` method (`@colony/colony-js-client`)
 
 **Maintenance**
 
 * Update `TokenClient` to use `DSToken` contract (`@colony/colony-js-client`)
+* Update `getPotBalance` to `getFundingPotBalance` (`@colony/colony-js-client`)
+* Update `PotAdded` to `FundingPotAdded` (`@colony/colony-js-client`)
 
 ## v1.10.0